// commerceos/libs/apps/products/src/lib/shared/interfaces/section.interface.ts
export interface BaseSection {
  images?: string[];
  price: number;
  salePrice: number;
  onSales: boolean;
  available: boolean;
  productType: ProductTypes;
  saleStartDate?: string;
  saleEndDate?: string;
}

export interface MainSection extends BaseSection {
  title: string;
  condition: ProductConditions;
  brand: string;
}